Rename blank.js to blankTemplate.js
