+++ src/iframe/src/blankTemplate.js
